+++ src/main/java/com/besson/arknights/block/ModAbstractContainerBE.java
@@ -2,10 +2,17 @@ package com.besson.arknights.block;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.ChestBlock;
 import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.block.entity.ChestBlockEntity;
 import net.minecraft.block.entity.LootableContainerBlockEntity;
+import net.minecraft.block.entity.ViewerCountManager;
+import net.minecraft.block.enums.ChestType;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.DoubleInventory;
 import net.minecraft.inventory.Inventories;
+import net.minecraft.inventory.Inventory;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.network.listener.ClientPlayPacketListener;
@@ -13,15 +20,49 @@ import net.minecraft.network.packet.Packet;
 import net.minecraft.network.packet.s2c.play.BlockEntityUpdateS2CPacket;
 import net.minecraft.screen.GenericContainerScreenHandler;
 import net.minecraft.screen.ScreenHandler;
+import net.minecraft.sound.SoundCategory;
+import net.minecraft.sound.SoundEvent;
+import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
+import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 public abstract class ModAbstractContainerBE extends LootableContainerBlockEntity {
     private DefaultedList<ItemStack> inv = createInventory();
+    private final ViewerCountManager stateManager = new ViewerCountManager() {
+
+        @Override
+        protected void onContainerOpen(World world, BlockPos pos, BlockState state) {
+            playSound(world, pos, state, SoundEvents.BLOCK_CHEST_OPEN);
+        }
+
+        @Override
+        protected void onContainerClose(World world, BlockPos pos, BlockState state) {
+            playSound(world, pos, state, SoundEvents.BLOCK_CHEST_CLOSE);
+        }
+
+        @Override
+        protected void onViewerCountUpdate(World world, BlockPos pos, BlockState state, int oldViewerCount, int newViewerCount) {
+            world.addSyncedBlockEvent(pos, state.getBlock(), 1, newViewerCount);
+        }
+
+        @Override
+        protected boolean isPlayerViewing(PlayerEntity player) {
+            if (!(player.currentScreenHandler instanceof GenericContainerScreenHandler)) {
+                return false;
+            } else {
+                Inventory inventory = ((GenericContainerScreenHandler)player.currentScreenHandler).getInventory();
+                return inventory == ModAbstractContainerBE.this;
+            }
+        }
+    };
+
     protected ModAbstractContainerBE(BlockEntityType<?> blockEntityType, BlockPos blockPos, BlockState blockState) {
         super(blockEntityType, blockPos, blockState);
     }
+
     public DefaultedList<ItemStack> createInventory() {
         return DefaultedList.ofSize(27, ItemStack.EMPTY);
     }
@@ -61,6 +102,21 @@ public abstract class ModAbstractContainerBE extends LootableContainerBlockEntit
             Inventories.writeNbt(nbt, this.inv);
         }
     }
+
+    @Override
+    public void onOpen(PlayerEntity player) {
+        if (!this.removed && !player.isSpectator()) {
+            this.stateManager.openContainer(player, this.getWorld(), this.getPos(), this.getCachedState());
+        }
+    }
+
+    @Override
+    public void onClose(PlayerEntity player) {
+        if (!this.removed && !player.isSpectator()) {
+            this.stateManager.closeContainer(player, this.getWorld(), this.getPos(), this.getCachedState());
+        }
+    }
+
     @Override
     public @Nullable Packet<ClientPlayPacketListener> toUpdatePacket() {
         return BlockEntityUpdateS2CPacket.create(this);
@@ -103,5 +159,12 @@ public abstract class ModAbstractContainerBE extends LootableContainerBlockEntit
     public void clear() {
         super.clear();
         markDirty();
+    }
+
+    static void playSound(World world, BlockPos pos, BlockState state, SoundEvent soundEvent) {
+        double d = (double)pos.getX() + 0.5;
+        double e = (double)pos.getY() + 0.5;
+        double f = (double)pos.getZ() + 0.5;
+        world.playSound(null, d, e, f, soundEvent, SoundCategory.BLOCKS, 0.5F, world.random.nextFloat() * 0.1F + 0.9F);
     }
 }